test/e2e/framework: split server code apart
