+++ test/e2e/conformance/webhook_test.go
@@ -45,6 +45,7 @@ import (
 	"github.com/kcp-dev/kcp/test/e2e/fixtures/wildwest/apis/wildwest/v1alpha1"
 	wildwestclientset "github.com/kcp-dev/kcp/test/e2e/fixtures/wildwest/client/clientset/versioned/cluster"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkserver "github.com/kcp-dev/kcp/test/e2e/framework/server"
 )
 
 func TestMutatingWebhookInWorkspace(t *testing.T) {
@@ -88,7 +89,7 @@ func TestMutatingWebhookInWorkspace(t *testing.T) {
 		Deserializer: deserializer,
 	}
 
-	port, err := framework.GetFreePort(t)
+	port, err := frameworkserver.GetFreePort(t)
 	require.NoError(t, err, "failed to get free port for test webhook")
 	dirPath := filepath.Dir(server.KubeconfigPath())
 	testWebhook.StartTLS(t, filepath.Join(dirPath, "apiserver.crt"), filepath.Join(dirPath, "apiserver.key"), port)
@@ -206,7 +207,7 @@ func TestValidatingWebhookInWorkspace(t *testing.T) {
 		Deserializer: deserializer,
 	}
 
-	port, err := framework.GetFreePort(t)
+	port, err := frameworkserver.GetFreePort(t)
 	require.NoError(t, err, "failed to get free port for test webhook")
 	dirPath := filepath.Dir(server.KubeconfigPath())
 	testWebhook.StartTLS(t, filepath.Join(dirPath, "apiserver.crt"), filepath.Join(dirPath, "apiserver.key"), port)
