+++ test/e2e/framework/server/port.go
@@ -1,0 +1,85 @@
+/*
+Copyright 2025 The KCP Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package server
+
+import (
+	"fmt"
+	"io"
+	"net"
+	"os"
+	"path/filepath"
+	"strconv"
+	"testing"
+)
+
+// GetFreePort asks the kernel for a free open port that is ready to use.
+func GetFreePort(t *testing.T) (string, error) {
+	t.Helper()
+
+	for {
+		addr, err := net.ResolveTCPAddr("tcp", "localhost:0")
+		if err != nil {
+			return "", fmt.Errorf("could not resolve free port: %w", err)
+		}
+
+		l, err := net.ListenTCP("tcp", addr)
+		if err != nil {
+			return "", fmt.Errorf("could not listen on free port: %w", err)
+		}
+		// This defer is in a loop, although it should only be retried a fixed number of times, and return.
+		defer func(c io.Closer) { //nolint:gocritic
+			if err := c.Close(); err != nil {
+				t.Errorf("could not close listener: %v", err)
+			}
+		}(l)
+		port := strconv.Itoa(l.Addr().(*net.TCPAddr).Port)
+		// Tests run in -parallel will run in separate processes, so we must use the file-system
+		// for sharing state and locking across them to coordinate who gets which port. Without
+		// some mechanism for sharing state, the following race is possible:
+		// - process A calls net.ListenTCP() to resolve a new port
+		// - process A calls l.Close() to close the listener to allow accessory to use it
+		// - process B calls net.ListenTCP() and resolves the same port
+		// - process A attempts to use the port, fails as it is in use
+		// Therefore, holding the listener open is our kernel-based lock for this process, and while
+		// we hold it open we must record our intent to disk.
+		lockDir := filepath.Join(os.TempDir(), "kcp-e2e-ports")
+		if err := os.MkdirAll(lockDir, 0755); err != nil {
+			return "", fmt.Errorf("could not create port lockfile dir: %w", err)
+		}
+		lockFile := filepath.Join(lockDir, port)
+		if _, err := os.Stat(lockFile); os.IsNotExist(err) {
+			// we've never seen this before, we can use it
+			f, err := os.Create(lockFile)
+			if err != nil {
+				return "", fmt.Errorf("could not record port lockfile: %w", err)
+			}
+			if err := f.Close(); err != nil {
+				return "", fmt.Errorf("could not close port lockfile: %w", err)
+			}
+			// the lifecycle of an accessory (and thereby its ports) is the test lifecycle
+			t.Cleanup(func() {
+				if err := os.Remove(lockFile); err != nil {
+					t.Errorf("failed to remove port lockfile: %v", err)
+				}
+			})
+			return port, nil
+		} else if err != nil {
+			return "", fmt.Errorf("could not access port lockfile: %w", err)
+		}
+		t.Logf("found a previously-seen port, retrying: %s", port)
+	}
+}
