+++ test/e2e/workspacetype/controller_test.go
@@ -40,6 +40,7 @@ import (
 	"github.com/kcp-dev/kcp/sdk/apis/third_party/conditions/util/conditions"
 	kcpclientset "github.com/kcp-dev/kcp/sdk/client/clientset/versioned/cluster"
 	"github.com/kcp-dev/kcp/test/e2e/framework"
+	frameworkserver "github.com/kcp-dev/kcp/test/e2e/framework/server"
 )
 
 func TestWorkspaceTypes(t *testing.T) {
@@ -47,7 +48,7 @@ func TestWorkspaceTypes(t *testing.T) {
 	framework.Suite(t, "control-plane")
 
 	type runningServer struct {
-		framework.RunningServer
+		frameworkserver.RunningServer
 		kcpClusterClient kcpclientset.ClusterInterface
 		orgPath          logicalcluster.Path
 	}
